move simulate command to tx sub-command, refactor a few things
