+++ x/msgfees/client/cli/query.go
@@ -4,14 +4,11 @@ import (
 	"context"
 	"encoding/base64"
 	"fmt"
-	"io/ioutil"
-	"os"
 
 	flag "github.com/spf13/pflag"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/provenance-io/provenance/x/msgfees/types"
 	"github.com/spf13/cobra"
 )
@@ -27,7 +24,6 @@ func GetQueryCmd() *cobra.Command {
 	}
 	queryCmd.AddCommand(
 		AllMsgBasedFeesCmd(),
-		GetCmdCalculateMsgBasedFees(),
 	)
 	return queryCmd
 }
@@ -66,61 +62,6 @@ func AllMsgBasedFeesCmd() *cobra.Command {
 	flags.AddPaginationFlagsToCmd(cmd, "markers")
 	flags.AddQueryFlagsToCmd(cmd)
 	return cmd
-}
-
-func GetCmdCalculateMsgBasedFees() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "simulate [msg_tx_json_file]",
-		Args:  cobra.ExactArgs(1),
-		Short: "Simulate transaction and return estimated costs with possible msg based fees.",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			// sdk.Tx
-			theTx, err := ReadTxFromFile(clientCtx, args[0])
-			if err != nil {
-				return err
-			}
-			txBytes, err := clientCtx.TxConfig.TxEncoder()(theTx)
-			if err != nil {
-				return err
-			}
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			var response *types.CalculateTxFeesResponse
-			if response, err = queryClient.CalculateTxFees(
-				context.Background(),
-				&types.CalculateTxFeesRequest{TxBytes: txBytes},
-			); err != nil {
-				fmt.Printf("failed to calculate fees: %s\n", err.Error())
-				return nil
-			}
-			return clientCtx.PrintProto(response)
-		},
-	}
-	flags.AddTxFlagsToCmd(cmd)
-	return cmd
-}
-
-// Read and decode a StdTx from the given filename.  Can pass "-" to read from stdin.
-func ReadTxFromFile(ctx client.Context, filename string) (tx sdk.Tx, err error) {
-	var bytes []byte
-
-	if filename == "-" {
-		bytes, err = ioutil.ReadAll(os.Stdin)
-	} else {
-		bytes, err = ioutil.ReadFile(filename)
-	}
-
-	if err != nil {
-		return
-	}
-
-	return ctx.TxConfig.TxJSONDecoder()(bytes)
 }
 
 // sdk ReadPageRequest expects binary but we encoded to base64 in our marshaller
